Use a QueryBuilder to build queries. Add PokemonColumn.
A QueryBuilder abstracts a "current" mapped class, no matter how deep we
are in the "tree" of joined tables.
It also prevents joining a single relation multiple times.

PokemonColumn is currently a trivial subclass of ForeignColumn, but
in the future it's meant to be more transparent, so the Pokemon/PokemonForm
split stays hidden.